fix(binder): PG smallest-fit integer literal typing + numeric-widening overload pass
Audit P1: every whole-number literal was typed bigint and overload
resolution required an EXACT signature string match, so fn(42) against a
multi-overload function with fn(integer) reported a false "no function
matches" (the sole-overload fallback only saved single-overload functions).

- LiteralType now follows PG's smallest-fit rule: integer when the value
  fits int4, bigint when it fits int8, numeric beyond (radix forms and
  PG16 '_' separators handled). The prior SemanticModelTests pin of
  42 → bigint updated to the correct verdict.
- BindFuncCall gains a numeric-widening pass after exact match fails:
  a call arg widens along PG's implicit chain smallint → integer → bigint
  → numeric → real → double precision; unknown-typed args are wildcards;
  the candidate is accepted only when UNAMBIGUOUS (two survivors = stay
  unresolved, conservative). Signatures split on top-level commas only, so
  numeric(10,2)-style parameter types don't shear the arg list.

Verified: 3 new OverloadWideningTests; full Core suite 25,253/0 with live PG18.

+++ tests/PgProj.Core.Tests/SemanticModelTests.cs
@@ -196,7 +196,7 @@ public sealed class SemanticModelTests
 
         var items = bound.SelectItems;
         Assert.Equal("text", items[0].Type.Name);
-        Assert.Equal("bigint", items[1].Type.Name);
+        Assert.Equal("integer", items[1].Type.Name);     // PG smallest-fit: 42 is int4, not bigint
         Assert.Equal("boolean", items[2].Type.Name);     // comparison → boolean
     }
